Edit access to some things, mostly for add-ons (#3)
* fix FenceGate

* access change to avoid reflections on add-ons

* access to Altar Entry and methods for add-ons

* access to categories for add-ons

* access to the water state of the crucible

This is a bit more personal, it's for the Urn that fills them. As always to reduce the obfuscation calls.

* access to startBurning() for add-ons

As the jar, it's just for me to avoid reflections. Your choice to discard or accept this.

* Remove Forge license.

LICENSE.txt describes the Minecraft Forge license, not the license for the Eidolon project, so I am removing it from this repository.

Co-authored-by: Alexthw <77152778+Alexthw46@users.noreply.github.com>
Co-authored-by: elucent <9532786+elucent@users.noreply.github.com>